remove redundant codes, reorganize model codes
